Use existing virtual environment and Pipenv

+++ bootstrap.sh
@@ -3,14 +3,23 @@
 set -euo pipefail
 
 export PIPENV_VERBOSITY=-1
+type pipenv > /dev/null && PIPENV_EXISTS=1
 
-echo "### Creating virtual environment…\c"
-python3 -m venv --system-site-packages .venv
-echo "done! Entering virtual environment."
+if [ ! -f .venv/bin/activate ]; then
+    echo "### Creating virtual environment…\c"
+    python3 -m venv --system-site-packages .venv
+    echo "done!"
+else
+    echo "### Using existing virtual environment"
+fi
+
 . .venv/bin/activate
+python --version
 
-echo "\n### Installing Pipenv"
-pip install pipenv pip-autoremove
+if [ -z ${PIPENV_EXISTS} ]; then
+    echo "\n### Installing Pipenv"
+    pip install pipenv pip-autoremove
+fi
 
 echo "\n### Installing runtime dependencies"
 pipenv install
@@ -18,8 +27,11 @@ pipenv install
 echo "\n### Bootstrapping Dotfiles"
 pipenv run ansible-playbook ansible/dotfiles.yml
 
-echo "\n### Cleaning up"
-pip-autoremove pipenv pip-autoremove -y
+if [ -z ${PIPENV_EXISTS} ]; then
+    echo "\n### Cleaning up"
+    pip-autoremove pipenv pip-autoremove -y
+fi
+
 echo "\n### Exiting virtual environment…\c"
 deactivate
 echo "done!"
